+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFActionSheet.h
@@ -5,7 +5,7 @@
 //  Copyright (c) 2012-2013 Peter Steinberger. All rights reserved.
 //
 
-#import <UIKit/UIKit.h>
+#import "PSPDFKitGlobal.h"
 
 /// Helper to add block features to UIActionSheet.
 /// After block has been executed, it is set to nil, breaking potential retain cycles.