+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFRenderQueue.h
@@ -34,11 +34,18 @@ typedef NS_ENUM(NSUInteger, PSPDFRenderQueuePriority) {
 /// Requests a (freshly) rendered image from a specified document. Does not use the file cache.
 /// For options, see PSPDFPageRender.
 /// IF `queueAsNext` is set, the request will be processed ASAP, skipping the current queue.
-- (PSPDFRenderJob *)requestRenderedImageForDocument:(PSPDFDocument *)document forPage:(NSUInteger)page withSize:(CGSize)size clippedToRect:(CGRect)clipRect withAnnotations:(NSArray *)annotations options:(NSDictionary *)options priority:(PSPDFRenderQueuePriority)priority queueAsNext:(BOOL)queueAsNext delegate:(id<PSPDFRenderDelegate>)delegate;
+- (PSPDFRenderJob *)requestRenderedImageForDocument:(PSPDFDocument *)document andPage:(NSUInteger)page withSize:(CGSize)size clippedToRect:(CGRect)clipRect withAnnotations:(NSArray *)annotations options:(NSDictionary *)options priority:(PSPDFRenderQueuePriority)priority queueAsNext:(BOOL)queueAsNext delegate:(id<PSPDFRenderDelegate>)delegate;
 
 /// Cancel job.
 /// Use NSNotFound for `page` to delete all requests for the document.
-- (void)cancelRenderingForDocument:(PSPDFDocument *)document forPage:(NSUInteger)page delegate:(id<PSPDFRenderDelegate>)delegate async:(BOOL)async;
+- (void)cancelRenderingForDocument:(PSPDFDocument *)document andPage:(NSUInteger)page delegate:(id<PSPDFRenderDelegate>)delegate async:(BOOL)async;
+
+/// Return all queued jobs for the current `document` and `page`. (bound to `delegate`)
+- (NSArray *)renderJobsForDocument:(PSPDFDocument *)document andPage:(NSUInteger)page delegate:(id<PSPDFRenderDelegate>)delegate;
+
+/// Cancel a single render job.
+/// @return YES if cancellation was successful, NO if not found.
+- (BOOL)cancelJob:(PSPDFRenderJob *)job onlyIfQueued:(BOOL)onlyIfQueued;
 
 /// Cancel all queued and running jobs.
 - (void)cancelAllJobs;
@@ -64,9 +71,9 @@ typedef NS_ENUM(NSUInteger, PSPDFRenderQueuePriority) {
 
 @interface PSPDFRenderJob : NSObject
 
+@property (nonatomic, strong, readonly) PSPDFDocument *document;
 @property (nonatomic, assign, readonly) NSUInteger page;
-@property (nonatomic, weak,   readonly) PSPDFDocument *document;
-@property (nonatomic, assign, readonly) CGSize fullSize;
+@property (nonatomic, assign, readonly) CGSize size;
 @property (nonatomic, assign, readonly) CGRect clipRect;
 @property (nonatomic, assign, readonly) float zoomScale;
 @property (nonatomic, copy,   readonly) NSArray *annotations;